Simplify the `NimbusBridge` API to more closely match Android
This change completely hoists the responsibility of creating a
`WKWebView` and loading a URL out of Nimbus and into the consuming
app. At some point Nimbus may add back a helper of some sort, but for
now this drastically simplifies the API for `NimbusBridge` and gives
control to the apps where they need it.