Android: preserve and extend JS objects
To match the behavior on iOS, this change causes existing JS objects to
be preserved.

E.g. if the sample `DeviceExtension` is added to the NimbusBridge, the
following code should work, as it does on iOS
```js
(function() {

const DeviceExtension = window.DeviceExtension || (window.DeviceExtension = {})

DeviceExtension.doSomethingElse = () => {
  // Once Nimbus is initialized, window.DeviceExtension
  // will have { getDeviceInfo, doSomethingElse }
  DeviceExtension.getDeviceInfo().then(info => {
    console.log(info)
  })
}

})();
```

Prior to this change, the behavior depends on whether the above JS code
executes before or after Nimbus initialization.
With this change, both `getDeviceInfo` and `doSomethingElse` are
always available after initialization, regardless of initialization
order.