Add `Plugin` protocol with updated bind api
This protocol will replace `NimbusExtension` in v1.0. It takes an
additional parameter which is the bridge the plugin is being bound by.
There is a default implementation that calls through to the old method
for now, but this default implementation and the old protocol will be
removed in the near future.